Create/process reveal tx + notify contract

+++ src/contracts/qpi.h
@@ -2472,7 +2472,7 @@ namespace QPI
 	{
 		sint64 queryId;			///< ID of the oracle query that led to this notification.
 		uint32 subscriptionId;	///< ID of the oracle subscription or 0 in case of a pure oracle query.
-		uint8 status;			///< Oracle query status as defined in `network_messages/common_def.h`
+		uint32 status;			///< Oracle query status as defined in `network_messages/common_def.h`
 		typename OracleInterface::OracleReply reply;	///< Oracle reply if status == ORACLE_QUERY_STATUS_SUCCESS
 	};
